requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
configparser==4.0.2
coverage==5.0.3
crayons==0.3.0
defusedxml==0.6.0
Django==3.0.2
django-tastypie==0.14.3
djangorestframework==3.11.0
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
MarkupPy==1.14
model-bakery==1.0.2
odfpy==1.4.0
openpyxl==3.0.3
pip-check-reqs==2.0
python-dateutil==2.8.1
python-mimeparse==1.6.0
pytz==2019.3
PyYAML==5.3
requests==2.22.0
selenium==3.141.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.7
webdriver-manager==2.3.0
xlrd==1.2.0
xlwt==1.3.0